webui: Added tweaks section to Tools -> Other settings.
These are settings that are documented but not exposed on the webui,
either because they are low-level hacks for specific corner-cases,
experimental settings, or I just never had the chance to fully
integrate them into the webui.

Some of these will require a reboot, or any related service to be
manually restarted by the user.